requirements.txt
networkx>=1.11
pyyaml>=3.12